+++ test/dienstplan/api_test.clj
@@ -18,16 +18,31 @@
    [cheshire.core :as json]
    [clj-http.client :as http]
    [clojure.test :refer [deftest is testing use-fixtures]]
-   [dienstplan.core]
    [dienstplan.commands :as cmd]
-   [dienstplan.db]
+   [dienstplan.core]
+   [dienstplan.db :as db]
    [dienstplan.fixture :as fix]
-   [next.jdbc :as jdbc]
-   [dienstplan.db :as db]))
+   [dienstplan.helpers :as helpers]
+   [dienstplan.schedule :as schedule]
+   [next.jdbc :as jdbc])
+  (:import (java.time ZonedDateTime ZoneId)))
 
-;; Const
+;; Const & Helpers
 
+(def dt-before "2020-02-15T10:30:59.123000000-00:00")
 (def dt-now "2021-02-15T10:30:59.123000000-00:00")
+(def dt-after "2022-02-15T10:30:59.123000000-00:00")
+
+(defn zoned-dt
+  [dt-str]
+  (ZonedDateTime/parse dt-str))
+
+(defn inst->zoned-dt
+  [inst]
+  (.atZone inst (ZoneId/of "UTC")))
+
+(defn boom! [& _]
+  (throw (ex-info "Boom!" {:data nil})))
 
 (def events-request-base
   {:method :post
@@ -168,7 +183,7 @@
 (deftest ^:integration test-update-rota-ok
   (testing "Update existing rota, show details about rota"
     (with-redefs
-     [cmd/get-now-ts
+     [helpers/now-ts-sql
       (constantly (-> dt-now clojure.instant/read-instant-timestamp))]
       (let [create-response (http/request events-request-create)
             update-response
@@ -476,7 +491,7 @@
 (deftest ^:integration test-list-rota-ok
   (testing "List rota"
     (with-redefs
-     [cmd/get-now-ts
+     [helpers/now-ts-sql
       (constantly (-> dt-now clojure.instant/read-instant-timestamp))]
       (let [create-response-1 (http/request events-request-create)
             create-response-2
@@ -574,8 +589,7 @@
 (deftest ^:integration test-api-unhandled-exception
   (testing "API unhandled exception"
     (with-redefs
-     [cmd/text-trim (fn [& _]
-                      (throw (ex-info "Boom!" {:data nil})))]
+     [helpers/text-trim boom!]
       (let [request
             (merge
              events-request-base
@@ -595,3 +609,345 @@
              (-> response
                  :body
                  (json/parse-string true))))))))
+
+;; schedule
+
+(deftest ^:integration test-schedule-ok
+  (testing "Create new schedule"
+    (let [schedule-create-rotate-response
+          (http/request
+           (merge
+            events-request-base
+            {:body
+             (json/generate-string
+              {:event
+               {:text "<@U001> schedule create \"rotate my-rota\" 5 0 * * Mon-Fri"
+                :ts "1640250011.000100"
+                :team "T123"
+                :channel "C123"}})}))
+          schedule-create-duplicate-response
+          (http/request
+           (merge
+            events-request-base
+            {:body
+             (json/generate-string
+              {:event
+               {:text "<@U001> schedule create \"rotate my-rota\" 5 0 * * Mon-Fri"
+                :ts "1640250011.000100"
+                :team "T123"
+                :channel "C123"}})}))
+          schedule-create-who-response
+          (http/request
+           (merge
+            events-request-base
+            {:body
+             (json/generate-string
+              {:event
+               {:text "<@U001> schedule create \"who my-rota\" 0 9 * * Mon-Fri"
+                :ts "1640250011.000100"
+                :team "T123"
+                :channel "C123"}})}))
+          schedule-list-response
+          (http/request
+           (merge
+            events-request-base
+            {:body
+             (json/generate-string
+              {:event
+               {:text "<@U001> schedule list"
+                :ts "1640250011.000100"
+                :team "T123"
+                :channel "C123"}})}))
+          schedule-delete-who-response
+          (http/request
+           (merge
+            events-request-base
+            {:body
+             (json/generate-string
+              {:event
+               {:text "<@U001> schedule delete \"who my-rota\""
+                :ts "1640250011.000100"
+                :team "T123"
+                :channel "C123"}})}))
+          schedule-delete-rotate-response
+          (http/request
+           (merge
+            events-request-base
+            {:body
+             (json/generate-string
+              {:event
+               {:text "<@U001> schedule delete \"rotate my-rota\""
+                :ts "1640250011.000100"
+                :team "T123"
+                :channel "C123"}})}))
+          schedule-delete-non-existent-response
+          (http/request
+           (merge
+            events-request-base
+            {:body
+             (json/generate-string
+              {:event
+               {:text "<@U001> schedule delete \"shout my-rota\""
+                :ts "1640250011.000100"
+                :team "T123"
+                :channel "C123"}})}))
+          schedule-list-empty-response
+          (http/request
+           (merge
+            events-request-base
+            {:body
+             (json/generate-string
+              {:event
+               {:text "<@U001> schedule list"
+                :ts "1640250011.000100"
+                :team "T123"
+                :channel "C123"}})}))]
+      ;; create `rotate my-rota`
+      (is (= 200 (:status schedule-create-rotate-response)))
+      (is (=
+           {:channel "C123"
+            :text "Executable `rotate my-rota` successfully scheduled with `5 0 * * Mon-Fri`"}
+           (-> schedule-create-rotate-response
+               :body
+               (json/parse-string true))))
+
+      ;; duplicate: unique chanel-executable index violation
+      (is (= 200 (:status schedule-create-duplicate-response)))
+      (is (=
+           {:channel "C123"
+            :text "Duplicate schedule for `rotate my-rota` in the channel"}
+           (-> schedule-create-duplicate-response
+               :body
+               (json/parse-string true))))
+
+      ;; create `who my-rota`
+      (is (= 200 (:status schedule-create-who-response)))
+      (is (=
+           {:channel "C123"
+            :text "Executable `who my-rota` successfully scheduled with `0 9 * * Mon-Fri`"}
+           (-> schedule-create-who-response
+               :body
+               (json/parse-string true))))
+
+      ;; list schedules
+      (is (= 200 (:status schedule-list-response)))
+      (is (=
+           {:channel "C123"
+            :text "- `rotate my-rota` with `5 0 * * Mon-Fri`\n- `who my-rota` with `0 9 * * Mon-Fri`"}
+           (-> schedule-list-response
+               :body
+               (json/parse-string true))))
+
+      ;; delete who
+      (is (= 200 (:status schedule-delete-who-response)))
+      (is (=
+           {:channel "C123"
+            :text "Scheduling for `who my-rota` successfully deleted"}
+           (-> schedule-delete-who-response
+               :body
+               (json/parse-string true))))
+
+      ;; delete rotate
+      (is (= 200 (:status schedule-delete-rotate-response)))
+      (is (=
+           {:channel "C123"
+            :text "Scheduling for `rotate my-rota` successfully deleted"}
+           (-> schedule-delete-rotate-response
+               :body
+               (json/parse-string true))))
+
+      ;; delete non-existent
+      (is (= 200 (:status schedule-delete-non-existent-response)))
+      (is (=
+           {:channel "C123"
+            :text "Schedule not found for params: {:channel \"C123\", :executable \"shout my-rota\", :crontab nil, :run_at nil}"}
+           (-> schedule-delete-non-existent-response
+               :body
+               (json/parse-string true))))
+
+      ;; list is empty
+      (is (= 200 (:status schedule-list-empty-response)))
+      (is (=
+           {:channel "C123"
+            :text "No schedules found in the channel"}
+           (-> schedule-list-empty-response
+               :body
+               (json/parse-string true)))))))
+
+(deftest ^:integration test-schedule-runner-ok
+  (testing "Test background task for schedule processing"
+    (with-redefs
+      ;; `get-current-dt` used for parsing crontab when generating dt sequence
+     [org.pilosus.kairos/get-current-dt (constantly (zoned-dt dt-before))
+       ;; `now-ts-sql` used for `db/schedules-get` when selecting events
+      helpers/now-ts-sql (constantly (clojure.instant/read-instant-timestamp dt-now))]
+      (let [executable-str "rotate my-rota"
+            crontab-str "59 23 31 12 *"
+            command (format "<@U001> schedule create \"%s\" %s" executable-str crontab-str)
+            processed-events-before-scheduling (schedule/run nil)
+            _ (http/request events-request-create)
+            _ (http/request
+               (merge
+                events-request-base
+                {:body
+                 (json/generate-string
+                  {:event
+                   {:text command
+                    :ts "1640250011.000100"
+                    :team "T123"
+                    :channel "C123"}})}))
+            events (jdbc/with-transaction [conn db/db]
+                     (db/schedules-get
+                      conn
+                      (-> dt-after clojure.instant/read-instant-timestamp)))
+            processed-events-after-scheduling (schedule/run nil)]
+        (is (= 0 processed-events-before-scheduling))
+        (is (= 1 processed-events-after-scheduling))
+        (is (= 1 (count events)))
+        (let [{:keys [schedule/run_at
+                      schedule/crontab
+                      schedule/executable]} (first events)
+              zoned-run-at (inst->zoned-dt run_at)]
+          (is (= crontab-str crontab))
+          (is (= executable-str executable))
+          (is (= 59 (.getMinute zoned-run-at)))
+          (is (= 23 (.getHour zoned-run-at)))
+          (is (= 31 (.getDayOfMonth zoned-run-at)))
+          ;; extract int from the Month enum
+          (is (= 12 (-> zoned-run-at
+                        .getMonth
+                        .getValue)))
+          (is (= 2020 (.getYear zoned-run-at))))))))
+
+(deftest ^:integration test-schedule-runner-exception
+  (testing "Schedule processing failed"
+    (with-redefs
+     [db/schedule-update! boom!
+      org.pilosus.kairos/get-current-dt (constantly (zoned-dt dt-before))
+      helpers/now-ts-sql (constantly (clojure.instant/read-instant-timestamp dt-now))]
+      (let [command "<@U001> schedule create \"rotate my-rota\" 59 23 31 12 *"
+            response (http/request
+                      (merge
+                       events-request-base
+                       {:body
+                        (json/generate-string
+                         {:event
+                          {:text command
+                           :ts "1640250011.000100"
+                           :team "T123"
+                           :channel "C123"}})}))
+            events-before-processing
+            (jdbc/with-transaction [conn db/db]
+              (db/schedules-get
+               conn
+               (-> dt-after
+                   clojure.instant/read-instant-timestamp)))
+            processed-events (schedule/run nil)
+            events-after-processing
+            (jdbc/with-transaction [conn db/db]
+              (db/schedules-get
+               conn
+               (-> dt-after
+                   clojure.instant/read-instant-timestamp)))]
+        (is (= 1 (count events-before-processing)))
+        (is (= 1 processed-events))
+        (is (= 1 (count events-after-processing)))
+        (let [event-before (first events-before-processing)
+              event-after (first events-after-processing)]
+          (is (= event-before event-after)))))))
+
+(def schedule-invalid-args "Invalid arguments for `schedule` command: %s")
+
+(def params-schedule-command-invalid-args
+  [["<@U001> schedule create who my rota 0 9 * * Mon-Fri"
+    :executable
+    "Invalid executable, double quotes omitted"]
+   ["<@U001> schedule create \"who my rota\" Mon-Fri"
+    :crontab
+    "Invalid crontab"]])
+
+(deftest ^:integration test-schedule-command-invalid-args
+  (testing "Invalid `schedule` command arguments: "
+    (doseq [[command invalid-arg description] params-schedule-command-invalid-args]
+      (testing description
+        (let [response (http/request
+                        (merge
+                         events-request-base
+                         {:body
+                          (json/generate-string
+                           {:event
+                            {:text command
+                             :ts "1640250011.000100"
+                             :team "T123"
+                             :channel "C123"}})}))
+              actual (-> response
+                         :body
+                         (json/parse-string true)
+                         :text)
+              expected (cmd/fmt-schedule-invalid-arg invalid-arg)]
+          (is (= expected actual)))))))
+
+(deftest ^:integration test-schedule-command-duplicate
+  (testing "Duplicate schedule"
+    (let [executable "rotate my-rota"
+          crontab "0 9 * * Mon-Fri"
+          command (format "<@U001> schedule create \"%s\" %s"
+                          executable
+                          crontab)
+          request (merge
+                   events-request-base
+                   {:body
+                    (json/generate-string
+                     {:event
+                      {:text command
+                       :ts "1640250011.000100"
+                       :team "T123"
+                       :channel "C123"}})})
+          create-response (http/request request)
+          duplicate-response (http/request request)
+          created (-> create-response
+                      :body
+                      (json/parse-string true)
+                      :text)
+          duplicate (-> duplicate-response
+                        :body
+                        (json/parse-string true)
+                        :text)]
+      (is (= created (format "Executable `%s` successfully scheduled with `%s`"
+                             executable crontab)))
+      (is (= duplicate (format "Duplicate schedule for `%s` in the channel"
+                               executable))))))
+
+(def params-schedule-command-db-exceptions
+  [["<@U001> schedule create \"rotate my-rota\" 0 9 * * Mon-Fri"
+    "Boom!"
+    "Error on schedule insert"]
+   ["<@U001> schedule delete \"rotate my-rota\""
+    "Boom!"
+    "Error on schedule delete"]
+   ["<@U001> schedule list"
+    "Boom!"
+    "Error on listing schedules"]])
+
+(deftest ^:integration test-schedule-command-db-exceptions
+  (testing "DB exception for schedule subcommand: "
+    (with-redefs [next.jdbc.sql/insert! boom!
+                  next.jdbc.sql/delete! boom!
+                  next.jdbc/execute! boom!]
+      (doseq [[command expected description] params-schedule-command-db-exceptions]
+        (testing description
+          (let [request (merge
+                         events-request-base
+                         {:body
+                          (json/generate-string
+                           {:event
+                            {:text command
+                             :ts "1640250011.000100"
+                             :team "T123"
+                             :channel "C123"}})})
+                response (http/request request)
+                actual (-> response
+                           :body
+                           (json/parse-string true)
+                           :text)]
+            (is (= expected actual))))))))