+++ CHANGELOG.md
@@ -6,6 +6,19 @@ file. This change log follows the conventions of
 
 ## [Unreleased]
 
+## [1.1.89] - 2023-07-31
+
+### Added
+
+- `schedule` command with
+  [vixie-cron](https://man7.org/linux/man-pages/man5/crontab.5.html)
+  format for scheduling
+  ([#55](https://github.com/pilosus/dienstplan/issues/55))
+
+### Changed
+
+- Log level highlighting for plain text logging
+
 ## [1.0.83] - 2023-07-29
 
 ### Fixed
@@ -230,7 +243,8 @@ documentation website added. Project has made it to the version 1.0.0!
 ### Added
 - Bot app MVP
 
-[Unreleased]: https://github.com/pilosus/dienstplan/compare/1.0.83...HEAD
+[Unreleased]: https://github.com/pilosus/dienstplan/compare/1.1.89...HEAD
+[1.1.89]: https://github.com/pilosus/dienstplan/compare/1.0.83...1.1.89
 [1.0.83]: https://github.com/pilosus/dienstplan/compare/1.0.82...1.0.83
 [1.0.82]: https://github.com/pilosus/dienstplan/compare/1.0.81...1.0.82
 [1.0.81]: https://github.com/pilosus/dienstplan/compare/1.0.0...1.0.81